Add files via upload
The stock market has consistently proven to be a good place to invest in and save for the
future.
There are a lot of compelling reasons to invest in stocks like it can help in fighting inflation,
create wealth, and also provides some tax benefits. It is important to maintain a diversified
portfolio when investing in stocks in order to maximize earnings under any market condition.
By doing a cluster analysis, one can identify stocks that exhibit similar characteristics and ones
which exhibit minimum correlation. This will help investors better analyze stocks across
different market segments and help protect against risks that could make the portfolio
vulnerable to losses. By analyzing the data, grouping the stocks based on the attributes
provided, and sharing insights about the characteristics of each group.